% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{Create a ggheat object}
\usage{
ggheat(data, id.vars = c(), distfun = dist, hclustfun = hclust,
  rlabels = rownames(data), clabels = colnames(data), cold = TRUE,
  rowd = TRUE, ...)
}
\arguments{
\item{data}{heatmap data in wide format, possibly also including columns with contextual data}
}
\value{
a ggheat object
}
\description{
Create a ggheat object
}

