% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap-components.R
\name{heatmap_components}
\alias{heatmap_components}
\title{Create grid and dendro parts of a heatmap as separate ggplot objects}
\usage{
heatmap_components(x, rlabels = rownames(x), clabels = colnames(x),
  cold = TRUE, rowd = TRUE)
}
\arguments{
\item{x}{matrix of intensity values}

\item{rlabels}{a character vector of row labels with length equal to number of rows in x}

\item{clabels}{a character vector of column labels with length equal to number of columns in x}

\item{cold}{whether to cluster columns and produce a column dendro}

\item{rowd}{whether to cluster rows and produce a row dendro}
}
\value{
a named list with the following objects
\itemize{
 \item tile tiled intensity values
 \item rowd and cold row and column dendrograms
}\item transformed data used to plot tile, rowd and cold
}
\description{
Create grid and dendro parts of a heatmap as separate ggplot objects
}
\examples{
data(pms)

hmc <- heatmap_components(x = pms)

show_ggheatmap(hmc$tile,row_dendro = hmc$rowd,col_dendro = hmc$cold)
}

