% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-ggheatmap.R
\name{show_ggheatmap}
\alias{show_ggheatmap}
\title{Displays ggplot based heatmap composed of multiple ggplot objects}
\usage{
show_ggheatmap(tile, row_dendro = NULL, col_dendro = NULL, lstrip = NULL,
  bstrip = NULL, cold.height = 0.2, rowd.width = 0.2,
  lstrip.width = 0.05, bstrip.height = 0.05, laxis.width = 0.2,
  baxis.height = 0.1, legend.height = 0.1, preserve_scales = FALSE)
}
\arguments{
\item{tile}{ggplot object with a tiled grid of intensity values as produced by heatmap_components}

\item{row_dendro}{row dendrogram as produced by heatmap_components}

\item{col_dendro}{col dendromgram as produced by heatmap_components}

\item{lstrip}{list containing ggplots to display as strips along left axis}

\item{bstrip}{list containing ggplots to display as strips along bottom axis}

\item{cold.height}{relative height of column dendro}

\item{rowd.width}{relative width of row dendro}

\item{lstrip.width}{relative width of lstrip plots}

\item{bstrip.height}{relative height of bstrip plots}

\item{laxis.width}{relative width of left axis}

\item{baxis.height}{relative height of bottom axis}

\item{legend.height}{relative height of legend}

\item{preserve_scales}{Allows you to use a custom x scale. If TRUE the x scale will not be replaced. Use expand(c(0.0,0)) on your custom scale.}
}
\value{
a gtable object that can be displayed with grid.draw
}
\description{
Displays ggplot based heatmap composed of multiple ggplot objects
}
\examples{
data(pms)
hmc <- heatmap_components(x = pms)
show_ggheatmap(hmc$tile,row_dendro = hmc$rowd,col_dendro = hmc$cold)
}

