% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heat.R
\name{stat_heat}
\alias{stat_heat}
\title{Add a heatmap layer to a plot}
\usage{
stat_heat(mapping = NULL, data = NULL, geom = "tile",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Add a heatmap layer to a plot
}

